Change NewDocs to int64

+++ src/dbnode/storage/stats/query_stats_metrics.go
@@ -52,7 +52,7 @@ func DefaultQueryStatsTrackerForMetrics(opts instrument.Options) QueryStatsTrack
 
 func (t *queryStatsMetricsTracker) TrackStats(values QueryStatsValues) error {
 	t.recentDocs.Update(float64(values.RecentDocs))
-	t.totalDocs.Inc(int64(values.NewDocs))
+	t.totalDocs.Inc(values.NewDocs)
 	return nil
 }
